Extracted `NeroParser` from `Parser`.
